+++ doc_test.go
@@ -78,6 +78,18 @@ func ExampleMapSet_Missing() {
 	// Output: true false
 }
 
+func ExampleMapSet_Equal() {
+	k := slices.Values([]string{"b", "a", "b"})
+	fmt.Println(Set("a", "b").Equal(k), Set("a").Equal(k))
+	// Output: true false
+}
+
+func ExampleMapSet_IsSubset() {
+	k := slices.Values([]string{"b", "a", "b"})
+	fmt.Println(Set("a").IsSubset(k), Set("a", "c").IsSubset(k))
+	// Output: true false
+}
+
 func ExampleMapSet_IsSuperset() {
 	k := slices.Values([]string{"b", "a", "b"})
 	fmt.Println(Set("a", "b").IsSuperset(k), Set("a").IsSuperset(k))