+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 ### Added
+* `Equal` and `IsSubset`
 * `Collect` and `Insert`
 * `Union` and `SymmetricDifference`
 * `Memoize`